+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 - Enforce consistent ordering of keys in engine literal values which are JSON objects.
+- Made `handovers` on `incident_schedule.config.rotations.*` required, as it
+  was always required by the API, but was not marked as such in the provider.
 
 ## v5.9.1